requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohttp==3.9.5
aiopywttr==2.3.0
aiosignal==1.3.1
annotated-types==0.7.0
APScheduler==3.10.4
attrs==23.2.0
certifi==2024.7.4
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
numpy==2.0.1
pandas==2.2.2
pillow==10.4.0
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywttr-models==1.2.4
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
yarl==1.9.4
